Move pagestream_api to separate module
For general readability.
